multiplayer mostly working

+++ js/firestore.js
@@ -12,24 +12,27 @@ export async function init() {
   await auth.signInAnonymously();
 
   // Can't use query params because changing them causes a reload
-  if (window.location.hash.includes(",")) {
-    // Create new loop
-    const [bpm, beatCount] = window.location.hash.replace("#", "").split(",").map(val => +val);
+  let bpm, beatCount;
+  const createdNewLoop = window.location.hash.includes(",");
+  if (createdNewLoop) {
+    [bpm, beatCount] = window.location.hash.replace("#", "").split(",").map(val => +val);
     loopDocument = await loopsCollection.add({ bpm, beatCount });
     window.location.hash = '#' + loopDocument.id;
   } else {
     // Use existing loop
     loopDocument = await loopsCollection.doc(window.location.hash.replace("#", ""));
+    // Somehow nice unpacking syntax doesn't work
+    const data = (await loopDocument.get()).data()
+    bpm = data.bpm;
+    beatCount = data.beatCount;
   }
 
-  // Has bpm and beatCount attributes
-  return (await loopDocument.get()).data();
+  return { bpm, beatCount, createdNewLoop };
 }
 
-// TODO: take audio from track.channel instead of separate arg
 export async function addTrack(track) {
   if (track.firestoreId !== null) {
-    throw new Error("can't add track to db, already added?");
+    throw new Error("can't add track, already added?");
   }
 
   const trackDocument = await loopDocument.collection('tracks').add({
@@ -37,12 +40,38 @@ export async function addTrack(track) {
     audioBlob: firebase.firestore.Blob.fromUint8Array(new Uint8Array(track.channel.floatArray.buffer)),
     creator: auth.getUid(),
   });
-  console.log("Added track: " + trackDocument.id);
-
-  track.nameInput.addEventListener('input', () => trackDocument.update({ name: track.nameInput.value }));
   track.firestoreId = trackDocument.id;
+  console.log("Added track: " + trackDocument.id);
+}
+
+export async function onNameChanged(track, name) {
+  if (track.firestoreId === null) {
+    throw new Error("omg");
+  }
+  await loopDocument.collection('tracks').doc(track.firestoreId).update({ name });
 }
 
 export async function deleteTrack(track) {
+  if (track.firestoreId === null) {
+    throw new Error("can't delete track without knowing id");
+  }
   await loopDocument.collection('tracks').doc(track.firestoreId).delete()
+}
+
+export function addTracksChangedCallback(changeCallback) {
+  loopDocument.collection('tracks').onSnapshot(snapshot => {
+    // This checks whether the change event was created by current browser tab or not
+    // https://firebase.google.com/docs/firestore/query-data/listen#events-local-changes
+    if (!snapshot.hasPendingWrites) {
+      changeCallback(snapshot.docs.map(doc => {
+        const data = doc.data();
+        return {
+          id: doc.id,
+          floatArray: new Float32Array(data.audioBlob.toUint8Array().buffer),
+          name: data.name,
+          createdByCurrentUser: true,  // FIXME
+        };
+      }));
+    }
+  });
 }
